// src/app/shared/classes/insuranceBusiness.enum.ts
export enum InsuranceBusiness {
    BUPA = 30,
    BUPA_PR = 31,
    BANAMEX = 32,
    BANCAFE = 33,
    ASSURIA = 34,
    ALIANZA = 35,
    SEGUROS_HORIZONTE = 36,
    CITISEGUROS = 37,
    CITIBANK = 38,
    BUPA_ECUADOR = 39,
    BUPA_DR = 40,
    BUPA_MEXICO = 41,
    CLICO_BAHAMAS = 42,
    CLICO_TRINIDAD = 43,
    _UNKNOWN = 44,
    IHI = 45,
    RFG = 46,
    BUPA_BOLIVIA = 47,
    BRITISH_AMERICAN = 48,
    IHI_PSI = 49,
    IHI_BEACON = 50,
    IHI_PROSEGUROS = 51,
    IHI_HSBC = 52,
    RIMAC = 53,
    BUPA_BOLIVIA_IHI = 54,
    BUPA_GUATEMALA = 55,
    BUPA_PANAMA = 56,
    CRUZ_BLANCA = 57,
    BUPA_TRINIDAD_AND_TOBAGO = 58,
    SANITAS = 59,
    BUPA_COLOMBIA_SEGUROS_BOLIVAR = 60,
    BUPA_BRAZIL_CARE_PLUS = 61,
    BUPA_GLOBAL = 999,

}