// src/app/shared/services/provider-detail/entities/principalLanguage.dto.ts
/**
* PrincipalLanguageDto.ts
*
* @description: DTO principal language.
* @author Jose Daniel Hernandez M.
* @version 1.0
* @date 19-07-2020.
*
*/



/**
 * DTO principal language.
 */
export interface PrincipalLanguageDto {

  /**
   * languageKey
   */
  languageKey: string;

  /**
   * languageIsoCode
   */
  languageIsoCode: string;

  /**
   * name
   */
  name: string;

}